fixed int -> long conversino on movieLibrary

+++ Source/Plex.Library/ApiModels/Libraries/MovieLibrary.cs
@@ -46,7 +46,7 @@ namespace Plex.Library.ApiModels.Libraries
         /// <summary>
         ///
         /// </summary>
-        public int ContentChangedAt { get; set; }
+        public long ContentChangedAt { get; set; }
 
         /// <summary>
         ///
